[c-libs/game] Add instance ID tracking
All available IDs are added to a linked list (so it holds 65536 16-bit
elements at the start). When instances are created, the list shrinks by
the first element. That way, there's no searching for available IDs and
the operation can be done in a constant O(1) time. The only cost is
increased memory usage, but I've done some calculations and 65536
elements amounts to at best a couple of megabytes of memory, so a
negligible amount.

Also rewrote parts of the RND_gameInstanceSpawn function to check for
all error codes properly and report them to stderr.