[c-libs/game] Move inline function bodies to header
I've learned that it's actually required for inline functions to be
declared and implemented in the same file, so there we go.

I was a bit unsure if it should be defined in the header or source, but
then I found this post:
https://embeddedgurus.com/barr-code/2011/03/do-inline-function-bodies-belong-in-c-header-files/
it seems to suggest that in my situation it's better to have the
function in the header.