[c-libs/game] Fix initialization performance bottleneck
Initialization worked, but would take a few solid seconds to finish.
This was because of the RND_linkedListAdd function, which operates in
O(n) time due to the anatomy of linked lists. That means, when adding
65336 elements in succession, the time complexity is basically O(n^2),
which is bad. The fix is easy though, just remember the position of the
last element and append from there instead of starting from the finish
every time.