Update installation method, bitarray improvements
I removed the AUR package, because it caused some problems. I'm not
ruling out the possibility of bringing it back, but for the time being
stick to the manual installation method (more customizable).

Added special cases to bitarray bit operations, if both bitarrays are
aligned in memory with 8-bit blocks, instead of calculating everything
bit-by-bit, standard C bit operators are used to change entire blocks
at a time.