[queue] Fix undefined behavior in RND_queueCopy
The former behavior was, RND_queueCopy would allocate memory for the new
data block, and in case of cpy function's errors, it would terminate all
operation and return 3. This is a problem, because it prevents the user
from safely recovering from the error, as it's impossible to free the
new data block, because the "rest" of the array (that wasn't overwritten
from the random malloc state due to the termination) is populated with
random values.

This problem can be easily fixed by changing malloc to calloc and
initializing the new array with 0s. This way, it is possible to recover
by calling RND_queueDestroy with a dtor function capable of handling
(ignoring) NULL pointers.